Added json desc for machines, need more testing

+++ src/task.rs
@@ -2,9 +2,9 @@
 pub struct Task<'a> {
     pub id : &'a str,
 
-    pub mem_count: u32,
-    pub alu_count: u32,
-    pub fpu_count: u32,
+    pub mem_count: u64,
+    pub alu_count: u64,
+    pub fpu_count: u64,
 
     pub cache_miss: f32,
     pub l1_cache_miss: f32,
